refactor () : item -> lotto 로 패키지 이름 수정
