+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domain.item;
+package lotto.domain.lotto;
 
 import lotto.exception.ValidLottoException;
