+++ src/main/java/lotto/domain/lotto/LottoTickets.java
@@ -1,4 +1,4 @@
-package lotto.domain.item;
+package lotto.domain.lotto;
 
 import enums.Rank;
 import lotto.domain.Money;
