+++ src/test/java/lotto/domain/lotto/LottoTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.item;
+package lotto.domain.lotto;
 
 import lotto.exception.ValidLottoException;
 import org.junit.jupiter.api.BeforeEach;
