+++ src/test/java/lotto/domain/lotto/LottoTicketTest.java
@@ -1,4 +1,4 @@
-package lotto.domain.item;
+package lotto.domain.lotto;
 
 import enums.Rank;
 import org.junit.jupiter.api.BeforeEach;
